Syntax improvements

+++ modules/core/kotlin/com/huskerdev/grapl/core/display/impl/MacDisplayPeer.kt
@@ -11,9 +11,9 @@ class MacDisplayPeer(
         @JvmStatic private external fun nGetMainScreen(): Long
         @JvmStatic private external fun nGetScreens(): LongArray
 
-        @JvmStatic private external fun nGetSize(screen: Long): IntArray
-        @JvmStatic private external fun nGetPosition(screen: Long): IntArray
-        @JvmStatic private external fun nGetPhysicalSize(screen: Long): IntArray
+        @JvmStatic private external fun nGetSize(screen: Long): DoubleArray
+        @JvmStatic private external fun nGetPosition(screen: Long): DoubleArray
+        @JvmStatic private external fun nGetPhysicalSize(screen: Long): DoubleArray
         @JvmStatic private external fun nGetDpi(screen: Long): Double
         @JvmStatic private external fun nGetFrequency(screen: Long): Int
         @JvmStatic private external fun nGetName(screen: Long): String
@@ -39,7 +39,7 @@ class MacDisplayPeer(
     override val name = nGetName(handle)
     override val systemName = "Screen ${nGetIndex(handle)}"
 
-    private fun scaledSize(width: Int, height: Int): Size {
+    private fun scaledSize(width: Double, height: Double): Size {
         val dpi = dpi
         return Size((width * dpi).toInt(), (height * dpi).toInt())
     }
